feat(decompositions): Williamson
The Williamson decomposition is implemented in this patch.

`takagi.py` got renamed to `decompositions.py`, and `test_takagi.py` got
renamed to `test_decompositions.py`.

The `xp_symplectic_form` is defined for verifying that the
Williamson decomposition truly returns a symplectic matrix in the xp
basis. Moreover, the `is_symplectic` got generalized for usage with
multiple forms, depending on the basis.

The symplectic group-related stuff got moved to `piquasso._math.symplectic`.

+++ piquasso/_math/transformations.py
@@ -42,7 +42,7 @@ def quad_transformation(d):
         numpy.ndarray: The basis changing matrix.
     """
 
-    T = np.zeros((2 * d, 2 * d))
+    T = np.zeros((2 * d, 2 * d), dtype=int)
     for i in range(d):
         T[2 * i, i] = 1
         T[2 * i + 1, i + d] = 1
